feat: add Outlet logic for profile route
