+++ src/pages/profile-page.module.css
@@ -54,28 +54,3 @@
   opacity: .4;
   margin-top: 80px;
 }
-
-.form {
-  display: flex;
-  flex-direction: column;
-  row-gap: 24px;
-}
-
-.controlsBox {
-  display: flex;
-  justify-content: flex-end;
-  align-items: flex-start;
-  column-gap: 20px;
-}
-
-.resetButton {
-  padding: 16px 8px;
-}
-
-.submitButton {
-  padding: 16px 40px;
-}
-
-.profileInfoPending {
-  composes: text text_type_main-large from global;
-}
